requirements.txt
# === Core Flask ===
Flask==3.1.3
Werkzeug==3.1.6
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Flask-Login==0.6.3
Flask-Marshmallow==1.4.0
Flask-Limiter==3.8.0
Flask-Session==0.5.0

# === Serialización ===
marshmallow==4.2.2
marshmallow-sqlalchemy==1.4.2

# === Base de datos ===
PyMySQL==1.1.0
# cryptography: PyMySQL la usa para auth_socket/caching_sha2_password (MySQL 8+/MariaDB 11+).
# Range conservador para no arrastrar un major nuevo por accidente.
cryptography>=42.0.0,<46.0.0
pyodbc==5.3.0

# === API / Docs ===
flasgger==0.9.7.1
requests==2.32.0
Flask-CORS==4.0.0

# === Utilidades ===
python-dotenv==1.2.2
Pillow==12.1.1

# === Servidor producción ===
# Bump desde 21.2.0: fija CVE-2024-1135 (HTTP request smuggling).
gunicorn==23.0.0

# === Desarrollo / Tests ===
pytest==8.0.0
